requirements.txt
beautifulsoup4
requests
pandas
streamlit
schedule
matplotlib
seaborn
python-dotenv